# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sardar Bio Organic API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';

export interface AdminSession {
  access_token: string;
  user: AdminUser;
}