# lib/api-zod/src/generated/types/dealershipInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sardar Bio Organic API
 * OpenAPI spec version: 0.1.0
 */

export interface DealershipInput {
  name: string;
  firm_name: string;
  email: string;
  phone: string;
  gst_number: string;
  area_pincode: string;
  subject: string;
  message: string;
}