# lib/api-zod/src/generated/types/submissionStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sardar Bio Organic API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmissionStats {
  contact_total: number;
  dealership_total: number;
  recent_contact: number;
  recent_dealership: number;
}